OHRM5X-316:Add controller trait
